历史持仓交易记录

+++ wp/Model/OpenPositionModel.swift
@@ -23,8 +23,8 @@ class PositionModel: BaseModel {
     var closeTime:Int = 0       //平仓时间	(未平仓时为空)
     var closePrice:Int = 0      //平仓价格	未平仓时为当前价格
     var grossProfit:Int = 0     //交易盈亏	未平仓时为当前浮动盈亏
-    var limit:Int = 0           //止盈	不设则为空
-    var stop:Int = 0            //止损不设则为空
+    var limit:Double = 0           //止盈	不设则为空
+    var stop:Double = 0            //止损不设则为空
     var closeType:Int = 0       //平仓类型	1:普通平仓; 2:自动平仓; 3:爆仓平仓; 4:收盘平仓; 5:强制平仓
     var isDeferred:Int = 0      //是否过夜
     var deferred:Int = 0        //累计过夜费
